Lots of changes, mostly listed below:
ScenarioList is now DiscoveryList to save the keyword for a later idea.
Added a tips and objective option on startup
Cleaned up old if-else code, removed plenty of commented lines
Reserved locations for Forest and Village additions
Combined hunger and thirst into Nourishment
Rarity is now more functional and customizable on the fly
Reserved ExpGain in interfaces and Main.
Introduced Lifetime to measure how long a player's been alive.